Okay, Shape is nonexhaustive
